py-backend/requirements.txt
flask
flask-cors
python-dotenv
flask_sqlalchemy
Flask_Migrate
bcrypt
PyJWT
cloudinary
psycopg2-binary
flask_jwt_extended